[wip]feat:no need to set ip and password for sealer run (#1342)
* feat:no need to set ip and password for sealer run
